requirements.txt
alembic==1.13.3
aniso8601==9.0.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.0
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.5
